requirements.txt
pytest
requests~=2.32.3
allure-python-commons~=2.13.5